implement namespace stats
