+++ sqld/src/http/user/dump.rs
