+++ sqld/src/http/user/result_builder.rs
