+++ sqld/src/http/user/types.rs
