+++ sqld/src/stats.rs
@@ -1,19 +1,14 @@
-use std::fs::{File, OpenOptions};
-use std::io::Seek;
-use std::path::Path;
+use std::path::{Path, PathBuf};
 use std::sync::atomic::{AtomicU64, Ordering};
-use std::sync::Arc;
+use std::sync::{Arc, Weak};
 use std::time::Duration;
 
 use serde::{Deserialize, Serialize};
+use tokio::io::AsyncWriteExt;
+use tokio::task::JoinSet;
 
-#[derive(Clone, Default)]
+#[derive(Debug, Default, Serialize, Deserialize)]
 pub struct Stats {
-    inner: Arc<StatsInner>,
-}
-
-#[derive(Serialize, Deserialize, Default)]
-struct StatsInner {
     rows_written: AtomicU64,
     rows_read: AtomicU64,
     storage_bytes_used: AtomicU64,
@@ -22,70 +17,85 @@ struct StatsInner {
 }
 
 impl Stats {
-    pub fn new(db_path: &Path) -> anyhow::Result<Self> {
+    pub async fn new(
+        db_path: &Path,
+        join_set: &mut JoinSet<anyhow::Result<()>>,
+    ) -> anyhow::Result<Arc<Self>> {
         let stats_path = db_path.join("stats.json");
-        let stats_file = OpenOptions::new()
-            .read(true)
-            .write(true)
-            .create(true)
-            .open(stats_path)?;
+        let this = if stats_path.try_exists()? {
+            let data = tokio::fs::read_to_string(&stats_path).await?;
+            Arc::new(serde_json::from_str(&data)?)
+        } else {
+            Arc::new(Stats::default())
+        };
 
-        let stats_inner =
-            serde_json::from_reader(&stats_file).unwrap_or_else(|_| StatsInner::default());
-        let inner = Arc::new(stats_inner);
+        join_set.spawn(spawn_stats_persist_thread(
+            Arc::downgrade(&this),
+            stats_path.to_path_buf(),
+        ));
 
-        spawn_stats_persist_thread(inner.clone(), stats_file);
-
-        Ok(Self { inner })
+        Ok(this)
     }
 
     /// increments the number of written rows by n
     pub fn inc_rows_written(&self, n: u64) {
-        self.inner.rows_written.fetch_add(n, Ordering::Relaxed);
+        self.rows_written.fetch_add(n, Ordering::Relaxed);
     }
 
     /// increments the number of read rows by n
     pub fn inc_rows_read(&self, n: u64) {
-        self.inner.rows_read.fetch_add(n, Ordering::Relaxed);
+        self.rows_read.fetch_add(n, Ordering::Relaxed);
     }
 
     pub fn set_storage_bytes_used(&self, n: u64) {
-        self.inner.storage_bytes_used.store(n, Ordering::Relaxed);
+        self.storage_bytes_used.store(n, Ordering::Relaxed);
     }
 
     /// returns the total number of rows read since this database was created
     pub fn rows_read(&self) -> u64 {
-        self.inner.rows_read.load(Ordering::Relaxed)
+        self.rows_read.load(Ordering::Relaxed)
     }
 
     /// returns the total number of rows written since this database was created
     pub fn rows_written(&self) -> u64 {
-        self.inner.rows_written.load(Ordering::Relaxed)
+        self.rows_written.load(Ordering::Relaxed)
     }
 
     /// returns the total number of bytes used by the database (excluding uncheckpointed WAL entries)
     pub fn storage_bytes_used(&self) -> u64 {
-        self.inner.storage_bytes_used.load(Ordering::Relaxed)
+        self.storage_bytes_used.load(Ordering::Relaxed)
     }
 
     /// increments the number of the write requests which were delegated from a replica to primary
     pub fn inc_write_requests_delegated(&self) {
-        self.inner
-            .write_requests_delegated
+        self.write_requests_delegated
             .fetch_add(1, Ordering::Relaxed);
     }
 
     pub fn write_requests_delegated(&self) -> u64 {
-        self.inner.write_requests_delegated.load(Ordering::Relaxed)
+        self.write_requests_delegated.load(Ordering::Relaxed)
     }
 }
 
-fn spawn_stats_persist_thread(stats: Arc<StatsInner>, mut file: File) {
-    std::thread::spawn(move || loop {
-        if file.rewind().is_ok() {
-            file.set_len(0).unwrap();
-            let _ = serde_json::to_writer(&mut file, &stats);
+async fn spawn_stats_persist_thread(stats: Weak<Stats>, path: PathBuf) -> anyhow::Result<()> {
+    loop {
+        if let Err(e) = try_persist_stats(stats.clone(), &path).await {
+            tracing::error!("error persisting stats file: {e}");
         }
-        std::thread::sleep(Duration::from_secs(5));
-    });
+        tokio::time::sleep(Duration::from_secs(5)).await;
+    }
+}
+
+async fn try_persist_stats(stats: Weak<Stats>, path: &Path) -> anyhow::Result<()> {
+    let temp_path = path.with_extension("tmp");
+    let mut file = tokio::fs::OpenOptions::new()
+        .write(true)
+        .create(true)
+        .open(&temp_path)
+        .await?;
+    file.set_len(0).await?;
+    file.write_all(&serde_json::to_vec(&stats)?).await?;
+    file.flush().await?;
+    tokio::fs::rename(temp_path, path).await?;
+    Ok(())
 }
